Pull updated time out of table header for runs and favorites tab
